Simplify the component setup
Removes the intermediate component definitions which simplifies the file
count and the component keys in the system configuration.

One change of note is that the `topology-builder-fn` is now combined
with the processor. This creates a 1:1 relationship between the
processor and the topology. It's not a problem but something to be aware
of.

Signed-off-by: Ticean Bennett <ticean@reactioncommerce.com>