Processor mock testing example
Adds an example integration test with Kafka `TestTopologyDriver`. Mocking
facilities are available in the `jackdaw` Kafka/Kafka Streams library and in the
`rp.jackdaw` library that adds some extra mock wrappers.

I added an output topic to the topology to test that the testing works. But the pricewatch is still handled as a side effect and the original values are passed through unchanged. We obviously don't want to leave it in this condition but
make a big step toward  test coverage of the topology itself.