Stateful transformer for pricewatch matching
Adds a stateful `transform` to the topology. The `Transformer.transform`
function has access to the transformer state. It can access state stores. It can
also emit 0 or many messages downstream.

So, it can be used to emit one message per pricewatch that qualifies as a match.

This new topology emits such messages. The new messages contains both pricing info and the pricewatch data in their preliminary state. Expect changes to this schema.