Use a state-store for price watches
Adds a topic for storing price watches. This is an improvement over the
static Clojure atom that was hardcoded to the get the project
bootstrapped.

There's no interface for getting data into the topic. Any producer will
work. The Kafka CLI tools, Kafkacat or a GUI like Conduktor are good options.

The watch keys should be a composite key of `<product-id>:<watch-id>` so
that the records can be scanned by the Streams `range` reader. Such a
key will act as a product index. Given any product ID, the range can efficiently
scan and find all watches for the product.

TODO ---

Partitioning at the Watch Producer
There's no explicit partitioning happening on the watch messages, so
partitions will NOT be colocated!! As such the app won't find matches
properly. The watch records must be produced with a custom partition
that partitions on the product-id so that `watches` and `prices-by-id`
are colocated.

Find Watches by Range
This isn't in fully working condition yet, because the `range` lookup
isn't implemented yet. We're returning all records from the state store.
Expect that to explode.

Signed-off-by: Ticean Bennett <ticean@reactioncommerce.com>